Fix SSH connector compilation warnings.

+++ ssh/inc/com/centreon/connector/ssh/channel.hh
@@ -24,10 +24,10 @@
 # include <string>
 # include <time.h>
 
-namespace                  com {
-  namespace                centreon {
-    namespace              connector {
-      namespace            ssh {
+namespace                    com {
+  namespace                  centreon {
+    namespace                connector {
+      namespace              ssh {
         /**
          *  @class channel channel.hh "com/centreon/connector/ssh/channel.hh"
          *  @brief SSH channel used to execute a command.
@@ -35,37 +35,37 @@ namespace                  com {
          *  The channel class represents SSH channels used to execute
          *  commands on the remote host.
          */
-        class              channel {
+        class                channel {
          private:
-          enum             e_step {
+          enum               e_step {
             chan_open = 1,
             chan_exec,
             chan_read,
             chan_close
           };
-          LIBSSH2_CHANNEL* _channel;
-          std::string      _cmd;
-          unsigned long    _cmd_id;
-          LIBSSH2_SESSION* _session;
-          std::string      _stderr;
-          std::string      _stdout;
-          e_step           _step;
-          time_t           _timeout;
-                           channel(channel const& c);
-          channel&         operator=(channel const& c);
-          bool             _close();
-          bool             _exec();
-          bool             _open();
-          bool             _read();
+          LIBSSH2_CHANNEL*   _channel;
+          std::string        _cmd;
+          unsigned long long _cmd_id;
+          LIBSSH2_SESSION*   _session;
+          std::string        _stderr;
+          std::string        _stdout;
+          e_step             _step;
+          time_t             _timeout;
+                             channel(channel const& c);
+          channel&           operator=(channel const& c);
+          bool               _close();
+          bool               _exec();
+          bool               _open();
+          bool               _read();
 
          public:
-                           channel(LIBSSH2_SESSION* sess,
-                             std::string const& cmd,
-                             unsigned long cmd_id,
-                             time_t timeout);
-                           ~channel();
-          bool             run();
-          time_t           timeout() const;
+                             channel(LIBSSH2_SESSION* sess,
+                               std::string const& cmd,
+                               unsigned long cmd_id,
+                               time_t timeout);
+                             ~channel();
+          bool               run();
+          time_t             timeout() const;
         };
       }
     }
